[#3] feat: Common Modal->ModalContainer
