+++ src/components/Main/Main.tsx
@@ -1,7 +1,7 @@
-import React from 'react';
+import React, { useState } from 'react';
 import { datas } from 'data/main';
 import VideoWrapper from 'styles/mainStyles/videoComponents/videoWrapper';
-import Modal from 'common/Modal';
+import Modal from 'common/ModalContainer';
 import VideoModal from 'components/VideoModal';
 import LazyItem from './LazyItem';
 import VideoSelectBar from './VideoSelectBar';
@@ -13,15 +13,7 @@ const Main = () => {
   return (
     <div style={{ width: '65%', margin: 'auto' }}>
       <VideoSelectBar popularTags={datas.popularTags} />
-      <Modal
-        isOverlayOn
-        isPopup
-        width="50%"
-        height="50%"
-        borderRadius={5}
-        borderStyle="none"
-        contentComponent={<VideoModal />}
-      />
+      <Modal isPopup={false} contentComponent={<VideoModal />} />
       <VideoWrapper>
         <div>
           {lVideos.map((data) => (
